toolkit/components/passwordmgr/test/test_basic_form.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Login Manager</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: simple form fill

<script>
commonInit();
SimpleTest.waitForExplicitFinish();

/** Test for Login Manager: form fill, multiple forms. **/

function startTest() {
  is($_(1, "uname").value, "testuser", "Checking for filled username");
  is($_(1, "pword").value, "testpass", "Checking for filled password");

  SimpleTest.finish();
}

window.onload = startTest;
</script>

<p id="display"></p>

<div id="content" style="display: none">

  <form id="form1" action="formtest.js">
    <p>This is form 1.</p>
    <input  type="text"       name="uname">
    <input  type="password"   name="pword">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

</div>

<pre id="test"></pre>
</body>
</html>
